requirements.txt
boto3
beautifulsoup4
python-dotenv